requirements.txt
Django==5.1.1
django-bootstrap-pagination==1.7.1
django-countries==7.6.1
django-paypal==2.1
django-phonenumber-field==8.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
paypalrestsdk==1.13.3
pillow==10.4.0
stripe==11.1.0